Adicioanando coordinator ao projeto
