+++ MasterCoin/Source/Controller/ListCoin/ListCoinViewController.swift
@@ -33,19 +33,3 @@ class ListCoinViewController: UIViewController {
 
 }
 
-
-extension ListCoinViewController: UITableViewDelegate, UITableViewDataSource{
-  func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-    return 10
-  }
-  
-  func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-    let cell: ListCoinTableViewCell? = tableView.dequeueReusableCell(withIdentifier: ListCoinTableViewCell.identifier, for: indexPath) as? ListCoinTableViewCell
-    return cell ?? ListCoinTableViewCell()
-    
-  }
-}
-
-extension ListCoinViewController: UISearchBarDelegate{
-  
-}
